// src/app/models/clients/ClientDetCrudReq.model.ts
export interface ClientDetCrudReq{
    idClient: number;
    idApplication: number;
    idCompany: number;
    fecDocumentDate: Date;
    indFirstName: string;
    indSecondName: string;
    indNames: string;
    indNationality: string;
    indCountryBirth: string;
    fecBirthDate: Date;
    indEconomicAct: string;
    indRfc: string;
    indAddressStreet: string;
    indAddressNum: string;
    indAddressSuburb: string;
    indAddressCp: string;
    indAddressCity: string;
    indAddressTown: string;
    indAddressState: string;
    indPhoneNumber: string;
    indPhoneExt: string;
    indPhoneSmart: string;
    indCountry: string;
    indCurp: string;
    indEmail: string;
    indIdOficial: string;
    indIdName: string;
    indIdAuthority: string;
    indForeign: string;
    indFstreet: string;
    indFnumber: string;
    indFsuburb: string;
    indFcity: string;
    indFstate: string;
    indFcp: string;
    indFcountry: string;
    indBcontroller: string;
    indBcDoc: string;
    operationType: number;
    createdBy: string;
    indEstatus: string;
    username: string;
}